requirements.txt
asgiref==3.4.1
astroid==2.6.6
autopep8==1.5.7
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.9.6
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
